\name{RIVMgrowth-package}
\alias{RIVMgrowth-package}
\alias{RIVMgrowth}
\docType{package}
\title{
\packageTitle{RIVMgrowth}
}
\description{
\packageDescription{RIVMgrowth}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{RIVMgrowth}
\packageIndices{RIVMgrowth}
~~ An overview of how to use the package, including the most ~~
~~ important functions ~~
}
\author{
\packageAuthor{RIVMgrowth}

Maintainer: \packageMaintainer{RIVMgrowth}
}
\references{
~~ Literature or other references for background information ~~
}

\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
